fix(deepseek): share v4 thinking policy profile
