+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Docs: https://docs.openclaw.ai
 
 ### Changes
 
+- DeepSeek: expose the V4 `xhigh` and `max` thinking levels through the bundled provider policy surface, so fallback `/think` resolution matches the runtime provider hook. Fixes #77139. Thanks @Thiagocsoaresbh.
+- Gateway/startup: keep model-catalog test helpers and run-session lookup code out of the hot `server.impl` import graph, reducing default gateway benchmark readiness latency.
+- Channels/streaming: add unified `streaming.mode: "progress"` drafts with auto single-word status labels and shared progress configuration across Discord, Telegram, Matrix, Slack, and Microsoft Teams.
 - Gateway/startup: keep model-catalog test helpers and run-session lookup code out of the hot `server.impl` import graph, reducing default gateway benchmark readiness latency.
 - Channels/streaming: add unified `streaming.mode: "progress"` drafts with auto single-word status labels and shared progress configuration across Discord, Telegram, Matrix, Slack, and Microsoft Teams.
 - Slack/streaming: add `streaming.progress.render: "rich"` for Block Kit progress drafts backed by structured progress line data.
